Change CSipHasher's count variable to uint8_t
Summary:
> SipHash technically supports arbitrarily long inputs (at least, I couldn't find a limit in the paper), but only the low 8 bits of the length matter. Because of that we should use an unsigned type to track the length (as any signed type could overflow, which is UB). uint8_t is sufficient, however.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19931 | core#19931]]

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10241

+++ src/crypto/siphash.h
@@ -14,7 +14,8 @@ class CSipHasher {
 private:
     uint64_t v[4];
     uint64_t tmp;
-    int count;
+    // Only the low 8 bits of the input size matter.
+    uint8_t count;
 
 public:
     /** Construct a SipHash calculator initialized with 128-bit key (k0, k1) */
